requirements.txt
backports.shutil-get-terminal-size==1.0.0
cffi==1.6.0
decorator==4.0.9
Django==1.9.6
django-crispy-forms==1.6.0
django-extensions==1.6.7
enum34==1.1.6
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.1.0
pickleshare==0.7.2
pkg-resources==0.0.0
ptyprocess==0.5.1
pycparser==2.14
pygit2==0.24.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
